refactored navigation
