+++ client/components/NavBar.js
@@ -1,0 +1,52 @@
+import React from 'react';
+import { NavigationContainer, useNavigation } from '@react-navigation/native';
+import { Button, View, StyleSheet, Text, Image, Pressable, TouchableOpacity} from 'react-native';
+import Icon from 'react-native-vector-icons/FontAwesome';
+
+const NavBar = () => {
+    const navigation = useNavigation();
+  return (
+        <View>
+        <View style={{ flexDirection: 'row', justifyContent: 'space-around', alignItems: 'center', backgroundColor: 'lightgray', padding: 10 }}>
+          {/* Navigation Buttons */}
+          <TouchableOpacity style={{ alignItems: 'center' }} onPress={() => navigation.navigate('Community')}>
+              <Icon name="home" size={20} color="black" alignItems="center"/>
+              <Text>Home</Text>
+          </TouchableOpacity>
+
+          <TouchableOpacity style={{ alignItems: 'center' }} onPress={() => navigation.navigate('Community')}>
+              <Icon name="calendar" size={20} color="black" alignItems="center"/>
+              <Text>Events</Text>
+          </TouchableOpacity>
+
+          <TouchableOpacity style={{ alignItems: 'center' }} onPress={() => navigation.navigate('Community')}>
+              <Icon name="search" size={20} color="black" alignItems="center"/>
+              <Text>Search</Text>
+          </TouchableOpacity>
+
+          <TouchableOpacity style={{ alignItems: 'center' }} onPress={() => navigation.navigate('Sign Up')}>
+              <Icon name="bookmark" size={20} color="black" alignItems="center"/>
+              <Text>Saved</Text>
+          </TouchableOpacity>
+
+          <TouchableOpacity style={{ alignItems: 'center' }} onPress={() => navigation.navigate('Sign Up')}>
+              <Icon name="shopping-cart" size={20} color="black" alignItems="center"/>
+              <Text>Shop</Text>
+          </TouchableOpacity>
+
+          <TouchableOpacity style={{ alignItems: 'center' }} onPress={() => navigation.navigate('Profile')}>
+              <Icon name="home" size={20} color="black" alignItems="center"/>
+              <Text>Profile</Text>
+          </TouchableOpacity>
+        </View>
+        <View style={{ alignItems: 'center', paddingBottom: 20 }}>
+          <Image
+            source={require('./ATLSFWlogo.jpg')}
+            style={{ width: 150, height: 50, resizeMode: 'contain' }}
+          />
+        </View>
+        </View>
+  );
+}
+
+export default NavBar;
